包括未更新到develop的增删改方法，和返回所有还未结束的活动。

+++ src/main/java/com/tencent/wxcloudrun/service/impl/ActivityServiceImpl.java
@@ -66,8 +66,14 @@ public class ActivityServiceImpl implements ActivityService {
     }
 
     @Override
+    public  Response<Activity> findDeletedActivity(Integer actID) {
+        Activity deletedActivity = activityMapper.findDeletedActivity(actID);
+        if (deletedActivity == null) return new Response<>();
+        return new Response<>(deletedActivity);
+    }
+
+    @Override
     public Response<Activity> getActivityDetails(Integer actID) {
-        System.out.println(actID);
         Activity activity = activityMapper.getActivityDetails(actID);
         if (activity != null) {
             return new Response<>(activity);
@@ -77,16 +83,73 @@ public class ActivityServiceImpl implements ActivityService {
     }
 
     @Override
+    public Response<List<Activity>> getOngoingActivities() {
+        List<Activity> ongoingActivities = activityMapper.getOngoingActivities();
+        return new Response<>(ongoingActivities);
+    }
+
+    @Override
     public Response<String> updateFullActivity(Integer actID, Activity updatedActivity) {
         if (updatedActivity.getImages() == null || updatedActivity.getImages().isEmpty()) {
             updatedActivity.setImagesJSON("{}");
         }
+
+        if (updatedActivity.getPayment() == null) updatedActivity.setPaymentJSON("{}");
+
+        Activity deletedActivity = activityMapper.findDeletedActivity(actID);
+        if (deletedActivity != null) {
+            if (updatedActivity.getTitle() == null || updatedActivity.getTitle().isEmpty()) {
+                updatedActivity.setTitle(deletedActivity.getTitle());
+            }
+
+            if (updatedActivity.getDescription() == null || updatedActivity.getDescription().isEmpty()) {
+                updatedActivity.setDescription(deletedActivity.getDescription());
+            }
+
+            if (updatedActivity.getStartDate() == null) {
+                updatedActivity.setStartDate(deletedActivity.getStartDate());
+            }
+
+            if (updatedActivity.getEndDate() == null) {
+                updatedActivity.setEndDate(deletedActivity.getEndDate());
+            }
+
+            if (updatedActivity.getDeadline() == null) {
+                updatedActivity.setDeadline(deletedActivity.getDeadline());
+            }
+
+            if (updatedActivity.getLocation() == null || updatedActivity.getLocation().isEmpty()) {
+                updatedActivity.setLocation(deletedActivity.getLocation());
+            }
+
+            if (updatedActivity.getImages() == null || updatedActivity.getImages().isEmpty()) {
+                if (deletedActivity.getImages() != null && !deletedActivity.getImages().isEmpty()) {
+                    updatedActivity.setImages(deletedActivity.getImages());
+                    updatedActivity.setImagesJSON(JSON.toJSONString(updatedActivity.getImages()));
+                }
+            }
+
+            if (updatedActivity.getPayment() == null && deletedActivity.getPayment() != null) {
+                updatedActivity.setPayment(deletedActivity.getPayment());
+                updatedActivity.setPaymentJSON(JSON.toJSONString(updatedActivity.getPayment()));
+            }
+
+            if (updatedActivity.getPrice() == null) updatedActivity.setPrice(deletedActivity.getPrice());
+
+            if (updatedActivity.getAdditionalInfoJSON() == null || updatedActivity.getAdditionalInfoJSON().isBlank()) {
+                updatedActivity.setAdditionalInfoJSON(deletedActivity.getAdditionalInfoJSON());
+            }
+        } else {
+            return new Response<>();
+        }
         activityMapper.updateFullActivity(actID, updatedActivity);
+
         return new Response<>("活动更新成功");
     }
 
     @Override
     public Response<String> deleteActivity(Integer actID) {
+        if (activityMapper.findDeletedActivity(actID) != null) return new Response<>();
         activityMapper.deleteActivity(actID);
         return new Response<>("活动删除成功");
     }
